:truck: Changed project file architecture
