+++ src/main/java/pvs/app/api/sonarqube/get/BugDTO.java
@@ -1,0 +1,14 @@
+package pvs.app.api.sonarqube.get;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+import java.util.Date;
+
+@Data
+@AllArgsConstructor
+public class BugDTO {
+    Date date;
+    Integer value;
+
+}
