+++ src/main/java/pvs/app/project/hyperlink/post/AddGithubRepositoryHyperlinkDTO.java
@@ -1,0 +1,9 @@
+package pvs.app.project.hyperlink.post;
+
+import lombok.Data;
+
+@Data
+public class AddGithubRepositoryHyperlinkDTO {
+    private Long projectId;
+    private String repositoryURL;
+}
