+++ src/main/java/pvs/app/project/hyperlink/post/AddSonarQubeHyperlinkDTO.java
@@ -1,0 +1,9 @@
+package pvs.app.project.hyperlink.post;
+
+import lombok.Data;
+
+@Data
+public class AddSonarQubeHyperlinkDTO {
+    private Long projectId;
+    private String repositoryURL;
+}
