--- src/main/java/pvs/app/project/repository/post/AddGithubRepositoryDTO.java
@@ -1,9 +1,0 @@
-package pvs.app.project.repository.post;
-
-import lombok.Data;
-
-@Data
-public class AddGithubRepositoryDTO {
-    private Long projectId;
-    private String repositoryURL;
-}
