--- src/main/java/pvs/app/project/repository/post/AddSonarRepositoryDTO.java
@@ -1,9 +1,0 @@
-package pvs.app.project.repository.post;
-
-import lombok.Data;
-
-@Data
-public class AddSonarRepositoryDTO {
-    private Long projectId;
-    private String repositoryURL;
-}
