--- src/main/java/pvs/app/sonar/api/get/BugDTO.java
@@ -1,14 +1,0 @@
-package pvs.app.sonar.api.get;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-import java.util.Date;
-
-@Data
-@AllArgsConstructor
-public class BugDTO {
-    Date date;
-    Integer value;
-
-}
