--- src/main/java/pvs/app/sonar/api/get/CodeCoverageDTO.java
@@ -1,14 +1,0 @@
-package pvs.app.sonar.api.get;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-import java.util.Date;
-
-@Data
-@AllArgsConstructor
-public class CodeCoverageDTO {
-    Date date;
-    Double value;
-
-}
