--- src/main/java/pvs/app/sonar/api/get/CodeSmellDTO.java
@@ -1,14 +1,0 @@
-package pvs.app.sonar.api.get;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-import java.util.Date;
-
-@Data
-@AllArgsConstructor
-public class CodeSmellDTO {
-    Date date;
-    Integer value;
-
-}
